% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plink_prs.R
\name{plink_prs}
\alias{plink_prs}
\title{Calculate PRS using plink2}
\usage{
plink_prs(
  base_df,
  id_col,
  effect_allele_col,
  effect_size_col,
  multithread = TRUE,
  pfile_dir,
  plink_bin
)
}
\arguments{
\item{base_df}{(dataframe) A dataframe containing at minimum variant ids, effect alleles, and corresponding effect sizes}

\item{id_col}{(character) Name of column containing variant ids}

\item{effect_allele_col}{(character) Name of column containing effect alleles}

\item{effect_size_col}{(character) Name of column containing effect estimates}

\item{multithread}{(logical) Should multi-threading be used to compute PRS across all pfiles in parallel (default = TRUE)}

\item{pfile_dir}{(character) Directory containing the plink files for the target dataset. If the directory contains multiple files (for example, one per chromosome), the PRS can be calculated in parallel.}

\item{plink_bin}{(character) Path to executable plink2 binary}
}
\value{
A \link[tibble:tibble-package]{tibble} containing rows for each individual in the target dataset, and a column containing their PRS
}
\description{
Calculate PRS using plink2
}
\examples{
\dontrun{
# load weights file into dataframe
weights_df <- fs::path_package("extdata", "PGS000667-PMBB.txt", package = "prsr") \%>\%
  vroom::vroom()

# calculate prs
prs_res <- plink_prs(
  base_df = weights_df,
  id_col = variant,
  effect_allele_col = effect_allele,
  effect_size_col = effect_weight_mg_dl,
  pfile_dir = "/project/PMBB/PMBB-Release-2020-2.0/Imputed/pgen/",
  plink_bin = "/scratch/Applications/PLINK/plink2_linux_x86_64_20211125/plink2a"
)
}
}
