% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prsr-package.R
\docType{package}
\name{prsr-package}
\alias{prsr}
\alias{prsr-package}
\title{prsr: Calculate PRS Using Plink}
\description{
A set of functions that use plink2 to calculate polygenic risk scores (PRS) on a target dataset given a base dataset that includes variants, effect alleles, and corresponding weights.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mglev1n/prsr}
  \item \url{https://mglev1n.github.io/prsr/}
  \item Report bugs at \url{https://github.com/mglev1n/prsr/issues}
}

}
\author{
\strong{Maintainer}: Michael Levin \email{michael.levin@pennmedicine.upenn.edu} (\href{https://orcid.org/0000-0002-9937-9932}{ORCID})

}
\keyword{internal}
